revert to manual priority setting; debug

+++ eth_priority_tests/types.h
@@ -56,7 +56,7 @@ struct ethernet_RX_frame
     char destination_mac[ETHER_ADDR_LEN];
     char source_mac[ETHER_ADDR_LEN];
     uint16_t data_size_or_type ; 
-    uint16_t align_buffer;
+    uint32_t align_buffer;
     union eth_payload payload;
 };
